fix: Move mock data access helper to test folder
